When multiple video streams are detected, don't drop (ignore) the other streams
This adds a nother section of '-c:# copy -map 0:#' commands for each video stream
other than the one selected to be reencoded, thus copying any other streams into
the new file. This is needed as some 'attachments' in mkv files (eg cover art) is
store as a video stream with the mjpeg codec.